chore: CR fixes & refactoring
