+++ backend/api/src/main/kotlin/io/tolgee/hateoas/ee/eeSubscription/EeSubscriptionModel.kt
@@ -25,6 +25,5 @@ open class EeSubscriptionModel(
   val status: SubscriptionStatus,
   val nonCommerical: Boolean,
   var lastValidCheck: Date?,
-  // TODO: TEST THIS
   val isPayAsYouGo: Boolean,
 ) : RepresentationModel<EeSubscriptionModel>(), Serializable
