+++ webapp/src/component/common/form/ResourceErrorComponent.tsx
@@ -18,7 +18,7 @@ export const ResourceErrorComponent: FunctionComponent<{
           .slice(0, props.limit)
           .map((e, index) => (
             <Box key={index}>
-              <Alert severity="error" data-cy={`error-message-${e}`}>
+              <Alert severity="error" data-cy="error-message">
                 <TranslatedError code={e} />
               </Alert>
             </Box>
